Fixed bug and got correct answer on 8_1. The issue was that we were counting multiple height-zero trees in a row next to the edge as visible. The puzzle only considered the first height 0 tree visible, onscuring the others.

+++ 08_1/treemap.go
@@ -86,8 +86,8 @@ func SetTreeVisibleFromEast(tree uint8, visible bool) uint8 {
 }
 
 // Gets the height of the tree.
-func GetTreeHeight(tree uint8) uint8 {
-	return tree & 0b_0000_1111
+func GetTreeHeight(tree uint8) int {
+	return int(tree & 0b_0000_1111)
 }
 
 // Returns whether a tree is visible from any direction.
